requirements.txt
amqp==5.0.2
asgiref==3.3.1
billiard==3.6.3.0
boto3
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
dj-database-url==0.5.0
Django==3.1.4
django-crispy-forms==1.10.0
django-extra-views==0.13.0
gunicorn==20.0.4
kombu==5.0.2
prompt-toolkit==3.0.8
psycopg2
python-lorem==1.1.2
pytz==2020.4
redis==3.5.3
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.2.0
